UP 3.4 WIP 2021-09-01

+++ trunk/Storebror/Ultrapack 3.4/src/com/maddox/il2/ai/TriggerMessage.java
@@ -9,7 +9,7 @@ import com.maddox.sas1946.il2.util.TrueRandom;
 class TriggerMessage extends Trigger {
 
     public TriggerMessage(String triggerName, int triggeredByArmy, int timeout, int posX, int posY, int radius, int altitudeMin, int altitudeMax, int triggeredBy, boolean triggerOnExit, int noObjectsMin, int probability, String linkActorName,
-            String displayMessage, int displayTime) {
+            String displayMessage, float displayTime) {
         super(triggerName, triggeredByArmy, timeout, posX, posY, radius, altitudeMin, altitudeMax, triggeredBy, triggerOnExit, noObjectsMin, probability, linkActorName, displayMessage, displayTime);
         this.setTriggerClass(TYPE_MESSAGE);
         if (this.getDisplayMessage() == "" || this.getDisplayMessage() == null) this.destroy();
